Crate gear_core

Gear core.

This library provides a runner for dealing with multiple little programs exchanging messages in a deterministic manner. To be used primary in Gear Substrate node implementation, but it is not limited to that.

Modules

code

Module for checked code.

costs

Costs module.

env

Environment for running a module.

gas

Gas module.

ids

Base identifiers for messaging primitives.

memory

Module for memory and allocations context.

message

Message processing module.

program

Module for programs.

Macros

charge_gas_token

We need this access as a macro because sometimes hiding the lifetimes behind a function won’t work out.