gear_core
Module buffer

Vector with limited len realization.

Structs

LimitedVec

Limited len vector. T is data type. E is overflow error type. N is max len which a vector can have.

RuntimeBufferSizeError

Runtime buffer size exceed error

Type Aliases

RuntimeBuffer

Buffer which size cannot be bigger then max allowed allocation size in runtime.