Enum gear_core::code::CodeError

pub enum CodeError {
    ImportSectionNotFound,
    MemoryEntryNotFound,
    ExportSectionNotFound,
    RequiredExportFnNotFound,
    NonGearExportFnFound,
    Decode,
    GasInjection,
    Encode,
    StartSectionIsFound,
}

Instrumentation error.

Variants

ImportSectionNotFound

The provided code doesn’t contain required import section.

MemoryEntryNotFound

The provided code doesn’t contain memory entry section.

ExportSectionNotFound

The provided code doesn’t contain export section.

RequiredExportFnNotFound

The provided code doesn’t contain the required init or handle export function.

NonGearExportFnFound

The provided code contains unnecessary function exports.

Decode

Error occurred during decoding original program code.

The provided code was a malformed Wasm bytecode or contained unsupported features (atomics, simd instructions, etc.).

GasInjection

Error occurred during injecting gas metering instructions.

This might be due to program contained unsupported/non-deterministic instructions (floats, manual memory grow, etc.).

Encode

Error occurred during encoding instrumented program.

The only possible reason for that might be OOM.

StartSectionIsFound

We restrict start sections in smart contracts.