Enum gear_core::code::SerializationError

pub enum SerializationError {

    UnexpectedEof,
    InvalidMagic,
    UnsupportedVersion(u32),
    InconsistentLength {
        expected: usize,
        actual: usize,
    },
    Other(&'static str),
    HeapOther(String),
    UnknownValueType(i8),
    UnknownBlockType(i32),
    UnknownTableElementType(i8),
    NonUtf8String,
    UnknownExternalKind(u8),
    UnknownInternalKind(u8),
    UnknownOpcode(u8),
    InvalidVarUint1(u8),
    InvalidVarInt32,
    InvalidVarInt64,
    InvalidVarUint32,
    InvalidVarUint64,
    InconsistentMetadata,
    InvalidSectionId(u8),
    SectionsOutOfOrder,
    DuplicatedSections(u8),
    InvalidMemoryReference(u8),
    InvalidTableReference(u8),
    InvalidLimitsFlags(u8),
    UnknownFunctionForm(u8),
    InvalidVarInt7(u8),
    InconsistentCode,
    InvalidSegmentFlags(u32),
    TooManyLocals,
    DuplicatedNameSubsections(u8),
    UnknownNameSubsectionType(u8),

}

Deserialization/serialization error

Variants

UnexpectedEof

Unexpected end of input.

InvalidMagic

Invalid magic.

UnsupportedVersion(u32)

Unsupported version.

InconsistentLength

Fields

expected: usize

Expected length of the definition.

actual: usize

Actual length of the definition.

Inconsistence between declared and actual length.

Other(&'static str)

Other static error.

HeapOther(String)

Other allocated error.

UnknownValueType(i8)

Invalid/unknown value type declaration.

UnknownBlockType(i32)

Invalid block type declaration.

UnknownTableElementType(i8)

Invalid/unknown table element type declaration.

NonUtf8String

Non-utf8 string.

UnknownExternalKind(u8)

Unknown external kind code.

UnknownInternalKind(u8)

Unknown internal kind code.

UnknownOpcode(u8)

Unknown opcode encountered.

InvalidVarUint1(u8)

Invalid VarUint1 value.

InvalidVarInt32

Invalid VarInt32 value.

InvalidVarInt64

Invalid VarInt64 value.

InvalidVarUint32

Invalid VarUint32 value.

InvalidVarUint64

Invalid VarUint64 value.

InconsistentMetadata

Inconsistent metadata.

InvalidSectionId(u8)

Invalid section id.

SectionsOutOfOrder

Sections are out of order.

DuplicatedSections(u8)

Duplicated sections.

InvalidMemoryReference(u8)

Invalid memory reference (should be 0).

InvalidTableReference(u8)

Invalid table reference (should be 0).

InvalidLimitsFlags(u8)

Invalid value used for flags in limits type.

UnknownFunctionForm(u8)

Unknown function form (should be 0x60).

InvalidVarInt7(u8)

Invalid varint7 (should be in -64..63 range).

InconsistentCode

Number of function body entries and signatures does not match.

InvalidSegmentFlags(u32)

Only flags 0, 1, and 2 are accepted on segments.

TooManyLocals

Sum of counts of locals is greater than 2^32.

DuplicatedNameSubsections(u8)

Duplicated name subsections.

UnknownNameSubsectionType(u8)

Unknown name subsection type.