Enum gear_core::costs::RuntimeCosts

pub enum RuntimeCosts {

    MeteringBlock(u32),
    Alloc,
    GasAvailable,
    MsgId,
    Origin,
    ProgramId,
    Source,
    Value,
    ValueAvailable,
    Size,
    Read(u32),
    BlockHeight,
    BlockTimestamp,
    SendInit,
    SendPush(u32),
    SendCommit(u32),
    ReplyCommit(u32),
    ReplyPush(u32),
    ReplyTo,
    Debug,
    ExitCode,
    Exit,
    Leave,
    Wait,
    Wake,
    CreateProgram(u32),

}

Enumerates syscalls that can be charged by gas meter.

Variants

MeteringBlock(u32)

Charge the gas meter with the cost of a metering block. The charged costs are the supplied cost of the block plus the overhead of the metering itself.

Alloc

Weight of calling alloc.

GasAvailable

Weight of calling gr_gas_available.

MsgId

Weight of calling gr_msg_id.

Origin

Weight of calling gr_origin.

ProgramId

Weight of calling gr_program_id.

Source

Weight of calling gr_source.

Value

Weight of calling gr_value.

ValueAvailable

Weight of calling gr_value_available.

Size

Weight of calling gr_size.

Read(u32)

Weight of calling gr_read.

BlockHeight

Weight of calling gr_block_height.

BlockTimestamp

Weight of calling gr_block_timestamp.

SendInit

Weight of calling gr_value_available.

SendPush(u32)

Weight of calling gr_send_push.

SendCommit(u32)

Weight of calling gr_send_commit.

ReplyCommit(u32)

Weight of calling gr_reply_commit.

ReplyPush(u32)

Weight of calling gr_reply_push.

ReplyTo

Weight of calling gr_reply_to.

Debug

Weight of calling gr_debug.

ExitCode

Weight of calling gr_exit_code.

Exit

Weight of calling gr_exit.

Leave

Weight of calling gr_leave.

Wait

Weight of calling gr_wait.

Wake

Weight of calling gr_wake.

CreateProgram(u32)

Weight of calling gr_create_program_wgas.

Implementations

impl RuntimeCosts

pub fn token(&self, s: &HostFnWeights) -> RuntimeToken

Returns a token with a weight given the parameters from HostFnWeights.