Module gear_core::env

Environment for running a module.

Structs

ClonedExtCarrier

ExtCarrier’s clone.

ExtCarrier

Struct for interacting with Ext.

ExtCarrierWithError

An error occurred during ExtCarrier::with_fallible which should be called only when inner is set

Enums

PageAction

Page access rights.

Traits

Ext

External api for managing memory, messages, allocations and gas-counting.