Trait gear_core::env::Ext

pub trait Ext {
    type Error: CoreError;

    fn alloc(
        &mut self,
        pages: WasmPageNumber,
        mem: &mut impl Memory
    ) -> Result<WasmPageNumber, Self::Error>;
    fn block_height(&mut self) -> Result<u32, Self::Error>;
    fn block_timestamp(&mut self) -> Result<u64, Self::Error>;
    fn origin(&mut self) -> Result<ProgramId, Self::Error>;
    fn send_init(&mut self) -> Result<usize, Self::Error>;
    fn send_push(
        &mut self,
        handle: usize,
        buffer: &[u8]
    ) -> Result<(), Self::Error>;
    fn send_commit(
        &mut self,
        handle: usize,
        msg: HandlePacket
    ) -> Result<MessageId, Self::Error>;
    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error>;
    fn reply_commit(
        &mut self,
        msg: ReplyPacket
    ) -> Result<MessageId, Self::Error>;
    fn reply_details(&mut self) -> Result<Option<ReplyDetails>, Self::Error>;
    fn source(&mut self) -> Result<ProgramId, Self::Error>;
    fn exit(&mut self) -> Result<(), Self::Error>;
    fn message_id(&mut self) -> Result<MessageId, Self::Error>;
    fn program_id(&mut self) -> Result<ProgramId, Self::Error>;
    fn free(&mut self, page: WasmPageNumber) -> Result<(), Self::Error>;
    fn debug(&mut self, data: &str) -> Result<(), Self::Error>;
    fn leave(&mut self) -> Result<(), Self::Error>;
    fn msg(&mut self) -> &[u8];
    fn gas(&mut self, amount: u32) -> Result<(), Self::Error>;
    fn charge_gas(&mut self, amount: u32) -> Result<(), Self::Error>;
    fn charge_gas_runtime(
        &mut self,
        costs: RuntimeCosts
    ) -> Result<(), Self::Error>;
    fn refund_gas(&mut self, amount: u32) -> Result<(), Self::Error>;
    fn gas_available(&mut self) -> Result<u64, Self::Error>;
    fn value(&mut self) -> Result<u128, Self::Error>;
    fn value_available(&mut self) -> Result<u128, Self::Error>;
    fn wait(&mut self) -> Result<(), Self::Error>;
    fn wake(&mut self, waker_id: MessageId) -> Result<(), Self::Error>;
    fn create_program(
        &mut self,
        packet: InitPacket
    ) -> Result<ProgramId, Self::Error>;
    fn forbidden_funcs(&self) -> &BTreeSet<&'static str>;

    fn send(&mut self, msg: HandlePacket) -> Result<MessageId, Self::Error> { ... }
    fn reply(&mut self, msg: ReplyPacket) -> Result<MessageId, Self::Error> { ... }

}

External api for managing memory, messages, allocations and gas-counting.

Required Associated Types

type Error: CoreError

An error issued in api

Required Methods

fn alloc(
&mut self,
pages: WasmPageNumber,
mem: &mut impl Memory
) -> Result<WasmPageNumber, Self::Error>

Allocate number of pages.

The resulting page number should point to pages consecutive memory pages.

fn block_height(&mut self) -> Result<u32, Self::Error>

Get the current block height.

fn block_timestamp(&mut self) -> Result<u64, Self::Error>

Get the current block timestamp.

fn origin(&mut self) -> Result<ProgramId, Self::Error>

Get the id of the user who initiated communication with blockchain, during which, currently processing message was created.

fn send_init(&mut self) -> Result<usize, Self::Error>

Initialize a new incomplete message for another program and return its handle.

fn send_push(&mut self, handle: usize, buffer: &[u8]) -> Result<(), Self::Error>

Push an extra buffer into message payload by handle.

fn send_commit(
&mut self,
handle: usize,
msg: HandlePacket
) -> Result<MessageId, Self::Error>

Complete message and send it to another program.

fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error>

Push an extra buffer into reply message.

fn reply_commit(&mut self, msg: ReplyPacket) -> Result<MessageId, Self::Error>

Complete reply message and send it to source program.

fn reply_details(&mut self) -> Result<Option<ReplyDetails>, Self::Error>

Read the message id, if current message is a reply.

fn source(&mut self) -> Result<ProgramId, Self::Error>

Get the source of the message currently being handled.

fn exit(&mut self) -> Result<(), Self::Error>

Terminate the program and transfer all available value to the address.

fn message_id(&mut self) -> Result<MessageId, Self::Error>

Get the id of the message currently being handled.

fn program_id(&mut self) -> Result<ProgramId, Self::Error>

Get the id of program itself

fn free(&mut self, page: WasmPageNumber) -> Result<(), Self::Error>

Free specific memory page.

Unlike traditional allocator, if multiple pages allocated via alloc, all pages should be free-d separately.

fn debug(&mut self, data: &str) -> Result<(), Self::Error>

Send debug message.

This should be no-op in release builds.

fn leave(&mut self) -> Result<(), Self::Error>

Interrupt the program, saving it’s state.

fn msg(&mut self) -> &[u8]

Access currently handled message payload.

fn gas(&mut self, amount: u32) -> Result<(), Self::Error>

Default gas host call.

fn charge_gas(&mut self, amount: u32) -> Result<(), Self::Error>

Charge some extra gas.

fn charge_gas_runtime(&mut self, costs: RuntimeCosts) -> Result<(), Self::Error>

Charge gas by RuntimeCosts token.

fn refund_gas(&mut self, amount: u32) -> Result<(), Self::Error>

Refund some gas.

fn gas_available(&mut self) -> Result<u64, Self::Error>

Tell how much gas is left in running context.

fn value(&mut self) -> Result<u128, Self::Error>

Value associated with message.

fn value_available(&mut self) -> Result<u128, Self::Error>

Tell how much value is left in running context.

fn wait(&mut self) -> Result<(), Self::Error>

Interrupt the program and reschedule execution.

fn wake(&mut self, waker_id: MessageId) -> Result<(), Self::Error>

Wake the waiting message and move it to the processing queue.

fn create_program(
&mut self,
packet: InitPacket
) -> Result<ProgramId, Self::Error>

Send init message to create a new program

fn forbidden_funcs(&self) -> &BTreeSet<&'static str>

Return the set of functions that are forbidden to be called.

Provided Methods

fn send(&mut self, msg: HandlePacket) -> Result<MessageId, Self::Error>

Send message to another program.

fn reply(&mut self, msg: ReplyPacket) -> Result<MessageId, Self::Error>

Produce reply to the current message.