gear_core
Module gas

Gas module.

Structs

GasAllowanceCounter

Gas allowance counter with some predefined maximum value.

GasAmount

Read-only representation of consumed GasCounter.

GasCounter

Gas counter with some predefined maximum gas.

GasInfo

The struct contains results of gas calculation required to process a message.

GasLeft

Gas limit and gas allowance left.

ValueCounter

Value counter with some predefined maximum value.

Enums

ChargeError

Charging error

ChargeResult

The result of charging gas.

CounterType

Enum representing current type of gas counter.

LockId

The id of the gas lock.

Traits

CountersOwner

Counters owner can change gas limit and allowance counters.