Module gear_core::gas_metering

This module is used to instrument a Wasm module with gas metering code.

Structs

ConstantCostRules

A type that implements Rules so that every instruction costs the same.

CustomConstantCostRules

This type provides the functionality of ConstantCostRules.

DbWeights

Describes DB access weights.

InstantiationWeights

Describes the weight for instantiation of the module.

InstructionWeights

Describes the weight for all categories of supported wasm instructions.

Limits

Describes the upper limits on various metrics.

MemoryWeights

Describes the weight for memory interaction.

RentWeights

Describes the weight for renting.

Schedule

Definition of the cost schedule and other parameterization for the wasm vm.

ScheduleRules

This type provides real gas cost of instructions on pallet-gear.

SyscallWeights

Describes the weight for each imported function that a program is allowed to call.

TaskWeights

Describes weights for running tasks.

Weight

Represents the computational time and storage space required for an operation.

Enums

MemoryGrowCost

Dynamic costs for memory growth.

Traits

Rules

An interface that describes instruction costs.

Functions

inject

Transforms a given module into one that charges gas for code to be executed by proxy of an imported gas metering function.

post_injection_handler

Helper procedure that makes adjustments after gas metering function injected.