Module gear_core::memory

Module for memory and allocations context.

Structs

AllocationsContext

Pages allocations context for the running program.

PageBuf

Buffer for gear page data.

PageNumber

Page number.

WasmPageNumber

Wasm page number.

Enums

Error

Memory error.

Traits

Memory

Backend wasm memory interface.

Functions

vec_page_data_map_to_page_buf_map

Tries to convert vector data map to page buffer data map. Makes it without buffer reallocations.

Type Definitions

HostPointer

Host pointer type. Host pointer can be 64bit or less, to support both we use u64.