Struct gear_core::memory::AllocationsContext

pub struct AllocationsContext { /* private fields */ }

Pages allocations context for the running program.

Implementations

impl AllocationsContext

pub fn new(
allocations: BTreeSet<WasmPageNumber>,
static_pages: WasmPageNumber,
max_pages: WasmPageNumber
) -> Self

New allocations context.

Provide currently running program_id, boxed memory abstraction and allocation manager. Also configurable static_pages and max_pages are set.

pub fn is_init_page(&self, page: WasmPageNumber) -> bool

Return true if the page is the initial page, it means that the page was already in the storage.

pub fn alloc(
&mut self,
pages: WasmPageNumber,
mem: &mut impl Memory
) -> Result<WasmPageNumber, Error>

Alloc specific number of pages for the currently running program.

pub fn free(&mut self, page: WasmPageNumber) -> Result<(), Error>

Free specific page.

Currently running program should own this page.

pub fn allocations(&self) -> &BTreeSet<WasmPageNumber>

Return reference to the allocation manager.

pub fn static_pages(&self) -> WasmPageNumber

Returns number of static pages.

pub fn into_parts(self) -> (BTreeSet<WasmPageNumber>, BTreeSet<WasmPageNumber>)

Decomposes this instance and returns allocations.