Trait gear_core::memory::Memory

pub trait Memory {
    fn grow(&mut self, pages: WasmPageNumber) -> Result<PageNumber, Error>;
    fn size(&self) -> WasmPageNumber;
    fn write(&mut self, offset: usize, buffer: &[u8]) -> Result<(), Error>;
    fn read(&self, offset: usize, buffer: &mut [u8]) -> Result<(), Error>;
    fn data_size(&self) -> usize;
    unsafe fn get_buffer_host_addr_unsafe(&self) -> HostPointer;

    fn get_buffer_host_addr(&self) -> Option<HostPointer> { ... }
}

Backend wasm memory interface.

Required Methods

fn grow(&mut self, pages: WasmPageNumber) -> Result<PageNumber, Error>

Grow memory by number of pages.

fn size(&self) -> WasmPageNumber

Return current size of the memory.

fn write(&mut self, offset: usize, buffer: &[u8]) -> Result<(), Error>

Set memory region at specific pointer.

fn read(&self, offset: usize, buffer: &mut [u8]) -> Result<(), Error>

Reads memory contents at the given offset into a buffer.

fn data_size(&self) -> usize

Returns the byte length of this memory.

unsafe fn get_buffer_host_addr_unsafe(&self) -> HostPointer

Get buffer addr unsafe.

Safety

if memory size is 0 then buffer addr can be garbage

Provided Methods

fn get_buffer_host_addr(&self) -> Option<HostPointer>

Returns native addr of wasm memory buffer in wasm executor