Trait gear_core::memory::PageU32Size

pub trait PageU32Size: Sized + Clone + Copy + PartialEq + Eq + PartialOrd + Ord {

    // Required methods
    fn size_non_zero() -> NonZeroU32;
    fn raw(&self) -> u32;
    unsafe fn new_unchecked(num: u32) -> Self;

    // Provided methods
    fn size() -> u32 { ... }
    fn from_offset(offset: u32) -> Self { ... }
    fn new(num: u32) -> Result<Self, PageError> { ... }
    fn offset(&self) -> u32 { ... }
    fn end_offset(&self) -> u32 { ... }
    fn to_page<PAGE: PageU32Size>(&self) -> PAGE { ... }
    fn add_raw(&self, raw: u32) -> Result<Self, PageError> { ... }
    fn sub_raw(&self, raw: u32) -> Result<Self, PageError> { ... }
    fn add(&self, other: Self) -> Result<Self, PageError> { ... }
    fn sub(&self, other: Self) -> Result<Self, PageError> { ... }
    fn inc(&self) -> Result<Self, PageError> { ... }
    fn dec(&self) -> Result<Self, PageError> { ... }
    fn align_down(&self, size: NonZeroU32) -> Self { ... }
    fn zero() -> Self { ... }
    fn iter_count(&self, count: Self) -> Result<PagesIter<Self>, PageError> { ... }
    fn iter_end(&self, end: Self) -> Result<PagesIter<Self>, PageError> { ... }
    fn iter_end_inclusive(
        &self,
        end: Self
    ) -> Result<PagesIterInclusive<Self>, PageError> { ... }
    fn iter_from_zero_inclusive(&self) -> PagesIterInclusive<Self>  { ... }
    fn iter_from_zero(&self) -> PagesIter<Self>  { ... }
    fn iter_once(&self) -> PagesIterInclusive<Self>  { ... }
    fn to_pages_iter<P: PageU32Size>(&self) -> PagesIterInclusive<P>  { ... }

}

Trait represents page with u32 size for u32 memory: max memory size is 2^32 bytes. All operations with page guarantees, that no addr or page number can be overflowed.

Required Methods

fn size_non_zero() -> NonZeroU32

Returns size of page.

fn raw(&self) -> u32

Returns raw page number.

unsafe fn new_unchecked(num: u32) -> Self

Constructs new page without any checks.

Safety

Doesn’t guarantee, that page offset or page end offset is in not overflowed.

Provided Methods

fn size() -> u32

Size as u32. Cannot be zero, because uses Self::size_non_zero.

fn from_offset(offset: u32) -> Self

Constructs new page from byte offset: returns page which contains this byte.

fn new(num: u32) -> Result<Self, PageError>

Constructs new page from raw page number with checks. Returns error if page will contain bytes, with offsets bigger then u32::MAX.

fn offset(&self) -> u32

Returns page zero byte offset.

fn end_offset(&self) -> u32

Returns page last byte offset.

fn to_page<PAGE: PageU32Size>(&self) -> PAGE

Returns new page, which impls PageU32Size, and contains self zero byte.

fn add_raw(&self, raw: u32) -> Result<Self, PageError>

Returns page which has number page.raw() + raw, with checks.

fn sub_raw(&self, raw: u32) -> Result<Self, PageError>

Returns page which has number page.raw() - raw, with checks.

fn add(&self, other: Self) -> Result<Self, PageError>

Returns page which has number page.raw() + other.raw(), with checks.

fn sub(&self, other: Self) -> Result<Self, PageError>

Returns page which has number page.raw() - other.raw(), with checks.

fn inc(&self) -> Result<Self, PageError>

Returns page which has number page.raw() + 1, with checks.

fn dec(&self) -> Result<Self, PageError>

Returns page which has number page.raw() - 1, with checks.

fn align_down(&self, size: NonZeroU32) -> Self

Aligns page zero byte and returns page which contains this byte. Normally if size % Self::size() == 0, then aligned byte is zero byte of the returned page.

fn zero() -> Self

Returns page, which has zero byte offset == 0.

fn iter_count(&self, count: Self) -> Result<PagesIter<Self>, PageError>

Returns iterator self..self + count.

fn iter_end(&self, end: Self) -> Result<PagesIter<Self>, PageError>

Returns iterator self..end.

fn iter_end_inclusive( &self, end: Self ) -> Result<PagesIterInclusive<Self>, PageError>

Returns iterator self..=end.

fn iter_from_zero_inclusive(&self) -> PagesIterInclusive<Self>

Returns iterator 0..=self

fn iter_from_zero(&self) -> PagesIter<Self>

Returns iterator 0..self

fn iter_once(&self) -> PagesIterInclusive<Self>

Returns iterator self..=self

fn to_pages_iter<P: PageU32Size>(&self) -> PagesIterInclusive<P>

To another page iterator. For example: PAGE1 has size 4 and PAGE2 has size 2:

Memory is splitted into PAGE1:
[<====><====><====><====><====>]
 0     1     2     3     4
Memory splitted into PAGE2:
[<=><=><=><=><=><=><=><=><=><=>]
 0  1  2  3  4  5  6  7  8  9
Then PAGE1 with number 2 contains [4, 5] pages of PAGE2,
and we returns iterator over [4, 5] PAGE2.