Module gear_core::message

Message processing module.

Structs

ContextOutcome

Context outcome.

ContextSettings

Context settings.

ContextStore

Store of previous message execution context.

Dispatch

Message with entry point.

HandleMessage

Message for Handle entry point. Represents a standard message that sends between actors.

HandlePacket

Handle message packet.

IncomingDispatch

Incoming message with entry point and previous execution context, if exists.

IncomingMessage

Incoming message.

InitMessage

Message for Init entry point. Used to initiate a newly created program.

InitPacket

Init message packet.

Message

An entity that is used for interaction between actors. Can transfer value and executes by programs in corresponding function: init, handle or handle_reply.

MessageContext

Context of currently processing incoming message.

ReplyDetails

Reply details data.

ReplyMessage

Message for Reply entry point. ReplyMessage is unique because of storing MessageId from message on what it replies, and can be the only one per some message execution.

ReplyPacket

Reply message packet.

StoredDispatch

Stored message with entry point and previous execution context, if exists.

StoredMessage

Stored message.

Enums

DispatchKind

Entry point for dispatch processing.

Traits

Packet

Message packet.

Type Definitions

ExitCode

Exit code type for message replies.

GasLimit

Gas limit type for message.

Payload

Payload type for message.

Salt

Salt type for init message.

Value

Value type for message.