Trait gear_core::message::WasmEntry

pub trait WasmEntry: Sized {
    // Required methods
    fn as_entry(&self) -> &str;
    fn try_from_entry(entry: &str) -> Option<Self>;

    // Provided method
    fn try_into_kind(&self) -> Option<DispatchKind> { ... }
}

Trait defining type could be used as entry point for a wasm module.

Required Methods

fn as_entry(&self) -> &str

Converting self into entry point name.

fn try_from_entry(entry: &str) -> Option<Self>

Converting entry point name into self object, if possible.

Provided Methods

fn try_into_kind(&self) -> Option<DispatchKind>

Tries to convert self into DispatchKind.