Module gear_core::pages

Module for memory pages.

Structs

GearPage

Page number.

PagesIter

U32 size pages iterator, to iterate continuously from one page to another.

PagesIterInclusive

U32 size pages iterator, to iterate continuously from one page to another, including the last one.

WasmPage

Wasm page number.

Enums

PageError

Errors when act with PageU32Size.

PageSizeNo

An enum which distinguishes between different page sizes.

Constants

GEAR_PAGE_SIZE

A gear page size, currently 4KiB to fit the most common native page size. This is size of memory data pages in storage. So, in lazy-pages, when program tries to access some memory interval - we can download just some number of gear pages instead of whole wasm page. The number of small pages, which must be downloaded, is depends on host native page size, so can vary.

WASM_PAGE_SIZE

A WebAssembly page has a constant size of 64KiB.

Traits

PageDynSize

Page with dynamic size.

PageNumber

Page number trait - page, which can return it number as u32.

PageU32Size

Trait represents page with u32 size for u32 memory: max memory size is 2^32 bytes. All operations with page guarantees, that no addr or page number can be overflowed.

SizeManager

Context where dynamic size pages store their sizes