gear_core

Module reservation

Gas reservation structures.

Structs

GasReservationSlot

Gas reservation slot.

GasReserver

Gas reserver.

ReservationNonce

An unchangeable wrapper over u64 value, which is required to be used as a “view-only” reservations nonce in a message execution context.

UnreservedReimbursement

Safety token returned when unreserved gas can be returned back to the gas counter.

Enums

GasReservationState

Gas reservation state.

Type Aliases

GasReservationMap

Gas reservations map.

GasReservationStates

Gas reservations states.