Type Alias gear_core::tasks::VaraScheduledTask

pub type VaraScheduledTask<AccountId> = ScheduledTask<AccountId, MessageId, bool>;

Alias for ScheduledTask used in vara-runtime, generic across AccountId used.

Aliased Type

enum VaraScheduledTask<AccountId> {
    PauseProgram(ActorId),
    RemoveCode(CodeId),
    RemoveFromMailbox(AccountId, MessageId),
    RemoveFromWaitlist(ActorId, MessageId),
    RemovePausedProgram(ActorId),
    WakeMessage(ActorId, MessageId),
    SendDispatch(MessageId),
    SendUserMessage {
        message_id: MessageId,
        to_mailbox: bool,
    },
    RemoveGasReservation(ActorId, ReservationId),
    RemoveResumeSession(u32),
}

Variants

PauseProgram(ActorId)

Pause program as out of rent one.

RemoveCode(CodeId)

Remove code from code storage as out of rent one.

RemoveFromMailbox(AccountId, MessageId)

Remove message from mailbox as out of rent one.

RemoveFromWaitlist(ActorId, MessageId)

Remove message from waitlist as out of rent one.

RemovePausedProgram(ActorId)

Remove paused program as dead one (issue #1014).

WakeMessage(ActorId, MessageId)

Delayed wake of the message at concrete block.

SendDispatch(MessageId)

Delayed message to program sending.

The message itself stored in DispatchStash.

SendUserMessage

Delayed message to user sending.

The message itself stored in DispatchStash.

Fields

message_id: MessageId

What message to send.

to_mailbox: bool

Should it be inserted into users mailbox.

RemoveGasReservation(ActorId, ReservationId)

Remove gas reservation.

RemoveResumeSession(u32)

๐Deprecated: Paused program storage was removed in pallet-gear-program

Remove resume program session.