Crate gear_core

Gear core.

This library provides a runner for dealing with multiple little programs exchanging messages in a deterministic manner. To be used primary in Gear Substrate node implementation, but it is not limited to that.

Modules

buffer

Vector with limited len realization.

code

Module for checked code.

costs

Costs module.

env

Environment for running a module.

env_vars

Execution settings

gas

Gas module.

ids

Base identifiers for messaging primitives.

memory

Module for memory and allocations context.

message

Message processing module.

pages

Module for memory pages.

percent

Basic struct for working with integer percentages.

program

Module for programs.

reservation

Gas reservation structures.

str

String with limited length implementation

Macros

declare_id

Declares data type for storing any kind of id for gear-core, which stores 32 bytes under the hood.