Crate gear_core

Gear core.

This library provides a runner for dealing with multiple little programs exchanging messages in a deterministic manner. To be used primary in Gear Substrate node implementation, but it is not limited to that.

Modules

buffer

Vector with limited len realization.

code

Module for checked code.

costs

Costs module.

env

Environment for running a module.

env_vars

Execution settings

gas

Gas module.

gas_metering

This module is used to instrument a Wasm module with gas metering code.

ids

Base identifiers for messaging primitives.

memory

Module for memory and allocations context.

message

Message processing module.

pages

Module for memory pages.

percent

Basic struct for working with integer percentages.

program

Module for programs.

reservation

Gas reservation structures.

rpc

This module contains types commonly used as output in RPC calls.

str

String with limited length implementation

tasks

The module provides primitives for all available regular or time-dependent tasks.

utils

Utility functions.