gear_core

Module buffer

Vector with limited len realization.

Structs

LimitedVec

Limited len vector. T is data type. E is overflow error type. N is max len which a vector can have.

PanicBuffer

Panic buffer which size cannot be bigger then max allowed payload size.

PayloadSizeError

Payload size exceed error

PayloadSlice

Wrapper for payload slice.

RuntimeBufferSizeError

Runtime buffer size exceed error

Constants

MAX_PAYLOAD_SIZE

Max payload size which one message can have (8 MiB).

Type Aliases

Payload

Payload type for message.

RuntimeBuffer

Buffer which size cannot be bigger then max allowed allocation size in runtime.