gear_core
Module costs

Costs module.

Structs

BlocksAmount

Chain blocks amount.

BytesAmount

Bytes amount.

CallsAmount

Some actions or calls amount.

CostOf

Gas cost per some type of action or data size.

ExtCosts

Execution externalities costs.

InstantiationCosts

Module instantiation costs.

IoCosts

IO costs.

LazyPagesCosts

Memory pages lazy access costs.

PagesCosts

Memory pages costs.

ProcessCosts

Costs for message processing

RentCosts

Holding in storages rent costs.

SyscallCosts

Program imported function call (syscall) costs.

Enums

CostToken

Enumerates syscalls that can be charged by gas meter.