Enum gear_core::costs::RuntimeCosts

pub enum RuntimeCosts {

    Alloc,
    Free,
    ReserveGas,
    UnreserveGas,
    GasAvailable,
    MsgId,
    Origin,
    ProgramId,
    Source,
    Value,
    ValueAvailable,
    Size,
    Read(u32),
    BlockHeight,
    BlockTimestamp,
    Random,
    SendInit,
    SendPush(u32),
    SendCommit(u32),
    ReplyCommit,
    ReplyPush(u32),
    ReplyTo,
    Debug(u32),
    ExitCode,
    Exit,
    Leave,
    Wait,
    WaitFor,
    WaitUpTo,
    Wake,
    CreateProgram(u32, u32),

}

Enumerates syscalls that can be charged by gas meter.

Variants

Alloc

Weight of calling alloc.

Free

Weight of calling free.

ReserveGas

Weight of calling gr_reserve_gas

UnreserveGas

Weight of calling gr_unreserve_gas

GasAvailable

Weight of calling gr_gas_available.

MsgId

Weight of calling gr_message_id.

Origin

Weight of calling gr_origin.

ProgramId

Weight of calling gr_program_id.

Source

Weight of calling gr_source.

Value

Weight of calling gr_value.

ValueAvailable

Weight of calling gr_value_available.

Size

Weight of calling gr_size.

Read(u32)

Weight of calling gr_read.

BlockHeight

Weight of calling gr_block_height.

BlockTimestamp

Weight of calling gr_block_timestamp.

Random

Weight of calling gr_random.

SendInit

Weight of calling gr_send_init.

SendPush(u32)

Weight of calling gr_send_push.

SendCommit(u32)

Weight of calling gr_send_commit.

ReplyCommit

Weight of calling gr_reply_commit.

ReplyPush(u32)

Weight of calling gr_reply_push.

ReplyTo

Weight of calling gr_reply_to.

Debug(u32)

Weight of calling gr_debug.

ExitCode

Weight of calling gr_exit_code.

Exit

Weight of calling gr_exit.

Leave

Weight of calling gr_leave.

Wait

Weight of calling gr_wait.

WaitFor

Weight of calling gr_wait_for.

WaitUpTo

Weight of calling gr_wait_up_to.

Wake

Weight of calling gr_wake.

CreateProgram(u32, u32)

Weight of calling gr_create_program_wgas.

Implementations

impl RuntimeCosts

pub fn token(&self, s: &HostFnWeights) -> RuntimeToken

Returns a token with a weight given the parameters from HostFnWeights.