gear_core
Module env

Environment for running a module.

Structs

DropPayloadLockBound

Result of calling a job within PayloadSliceLock::drop_with.

PayloadSliceAccess

A wrapper over mutable reference to PayloadSliceLock which can give to the caller the slice of the held payload.

PayloadSliceLock

Lock for the payload of the incoming/currently executing message.

UnlockPayloadBound

Result of calling Externalities::unlock_payload.

Traits

Externalities

External api and data for managing memory and messages, use by an executing program to trigger state transition in runtime.