Trait gear_core::env::Ext

pub trait Ext {
    type Error: CoreError;

    fn alloc(
        &mut self,
        pages: WasmPageNumber,
        mem: &mut impl Memory
    ) -> Result<WasmPageNumber, Self::Error>;
    fn block_height(&mut self) -> Result<u32, Self::Error>;
    fn block_timestamp(&mut self) -> Result<u64, Self::Error>;
    fn origin(&mut self) -> Result<ProgramId, Self::Error>;
    fn send_init(&mut self) -> Result<u32, Self::Error>;
    fn send_push(&mut self, handle: u32, buffer: &[u8]) -> Result<(), Self::Error>;
    fn send_commit(
        &mut self,
        handle: u32,
        msg: HandlePacket,
        delay: u32
    ) -> Result<MessageId, Self::Error>;
    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error>;
    fn reply_commit(
        &mut self,
        msg: ReplyPacket,
        delay: u32
    ) -> Result<MessageId, Self::Error>;
    fn reply_to(&mut self) -> Result<MessageId, Self::Error>;
    fn source(&mut self) -> Result<ProgramId, Self::Error>;
    fn exit(&mut self) -> Result<(), Self::Error>;
    fn exit_code(&mut self) -> Result<ExitCode, Self::Error>;
    fn message_id(&mut self) -> Result<MessageId, Self::Error>;
    fn program_id(&mut self) -> Result<ProgramId, Self::Error>;
    fn free(&mut self, page: WasmPageNumber) -> Result<(), Self::Error>;
    fn debug(&mut self, data: &str) -> Result<(), Self::Error>;
    fn leave(&mut self) -> Result<(), Self::Error>;
    fn read(&mut self) -> Result<&[u8], Self::Error>;
    fn size(&mut self) -> Result<usize, Self::Error>;
    fn random(&self) -> (&[u8], u32);
    fn charge_gas(&mut self, amount: u64) -> Result<(), Self::Error>;
    fn charge_gas_runtime(
        &mut self,
        costs: RuntimeCosts
    ) -> Result<(), Self::Error>;
    fn refund_gas(&mut self, amount: u64) -> Result<(), Self::Error>;
    fn reserve_gas(
        &mut self,
        amount: u64,
        duration: u32
    ) -> Result<ReservationId, Self::Error>;
    fn unreserve_gas(&mut self, id: ReservationId) -> Result<u64, Self::Error>;
    fn gas_available(&mut self) -> Result<u64, Self::Error>;
    fn value(&mut self) -> Result<u128, Self::Error>;
    fn value_available(&mut self) -> Result<u128, Self::Error>;
    fn wait(&mut self) -> Result<(), Self::Error>;
    fn wait_for(&mut self, duration: u32) -> Result<(), Self::Error>;
    fn wait_up_to(&mut self, duration: u32) -> Result<bool, Self::Error>;
    fn wake(&mut self, waker_id: MessageId, delay: u32) -> Result<(), Self::Error>;
    fn create_program(
        &mut self,
        packet: InitPacket,
        delay: u32
    ) -> Result<(MessageId, ProgramId), Self::Error>;
    fn forbidden_funcs(&self) -> &BTreeSet<&'static str>;
    fn counters(&self) -> (u64, u64);
    fn update_counters(&mut self, gas: u64, allowance: u64);
    fn out_of_gas(&mut self) -> Self::Error;
    fn out_of_allowance(&mut self) -> Self::Error;

    fn send(
        &mut self,
        msg: HandlePacket,
        delay: u32
    ) -> Result<MessageId, Self::Error> { ... }
    fn reply(
        &mut self,
        msg: ReplyPacket,
        delay: u32
    ) -> Result<MessageId, Self::Error> { ... }

}

External api for managing memory, messages, allocations and gas-counting.

Required Associated Types

type Error: CoreError

An error issued in api

Required Methods

fn alloc(
&mut self,
pages: WasmPageNumber,
mem: &mut impl Memory
) -> Result<WasmPageNumber, Self::Error>

Allocate number of pages.

The resulting page number should point to pages consecutive memory pages.

fn block_height(&mut self) -> Result<u32, Self::Error>

Get the current block height.

fn block_timestamp(&mut self) -> Result<u64, Self::Error>

Get the current block timestamp.

fn origin(&mut self) -> Result<ProgramId, Self::Error>

Get the id of the user who initiated communication with blockchain, during which, currently processing message was created.

fn send_init(&mut self) -> Result<u32, Self::Error>

Initialize a new incomplete message for another program and return its handle.

fn send_push(&mut self, handle: u32, buffer: &[u8]) -> Result<(), Self::Error>

Push an extra buffer into message payload by handle.

fn send_commit(
&mut self,
handle: u32,
msg: HandlePacket,
delay: u32
) -> Result<MessageId, Self::Error>

Complete message and send it to another program.

fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error>

Push an extra buffer into reply message.

fn reply_commit(
&mut self,
msg: ReplyPacket,
delay: u32
) -> Result<MessageId, Self::Error>

Complete reply message and send it to source program.

fn reply_to(&mut self) -> Result<MessageId, Self::Error>

Get the message id of the initial message.

fn source(&mut self) -> Result<ProgramId, Self::Error>

Get the source of the message currently being handled.

fn exit(&mut self) -> Result<(), Self::Error>

Terminate the program and transfer all available value to the address.

fn exit_code(&mut self) -> Result<ExitCode, Self::Error>

Get the exit code of the message being processed.

fn message_id(&mut self) -> Result<MessageId, Self::Error>

Get the id of the message currently being handled.

fn program_id(&mut self) -> Result<ProgramId, Self::Error>

Get the id of program itself

fn free(&mut self, page: WasmPageNumber) -> Result<(), Self::Error>

Free specific memory page.

Unlike traditional allocator, if multiple pages allocated via alloc, all pages should be free-d separately.

fn debug(&mut self, data: &str) -> Result<(), Self::Error>

Send debug message.

This should be no-op in release builds.

fn leave(&mut self) -> Result<(), Self::Error>

Interrupt the program, saving it’s state.

fn read(&mut self) -> Result<&[u8], Self::Error>

Access currently handled message payload.

fn size(&mut self) -> Result<usize, Self::Error>

Size of currently handled message payload.

fn random(&self) -> (&[u8], u32)

Returns a random seed for the current block with message id as a subject, along with the time in the past since when it was determinable by chain observers.

fn charge_gas(&mut self, amount: u64) -> Result<(), Self::Error>

Charge some extra gas.

fn charge_gas_runtime(&mut self, costs: RuntimeCosts) -> Result<(), Self::Error>

Charge gas by RuntimeCosts token.

fn refund_gas(&mut self, amount: u64) -> Result<(), Self::Error>

Refund some gas.

fn reserve_gas(
&mut self,
amount: u64,
duration: u32
) -> Result<ReservationId, Self::Error>

Reserve some gas for a few blocks.

fn unreserve_gas(&mut self, id: ReservationId) -> Result<u64, Self::Error>

Unreserve gas using reservation ID.

fn gas_available(&mut self) -> Result<u64, Self::Error>

Tell how much gas is left in running context.

fn value(&mut self) -> Result<u128, Self::Error>

Value associated with message.

fn value_available(&mut self) -> Result<u128, Self::Error>

Tell how much value is left in running context.

fn wait(&mut self) -> Result<(), Self::Error>

Interrupt the program and reschedule execution for maximum.

fn wait_for(&mut self, duration: u32) -> Result<(), Self::Error>

Interrupt the program and reschedule execution in duration.

fn wait_up_to(&mut self, duration: u32) -> Result<bool, Self::Error>

Interrupt the program and reschedule execution for maximum, but not more than duration.

fn wake(&mut self, waker_id: MessageId, delay: u32) -> Result<(), Self::Error>

Wake the waiting message and move it to the processing queue.

fn create_program(
&mut self,
packet: InitPacket,
delay: u32
) -> Result<(MessageId, ProgramId), Self::Error>

Send init message to create a new program

fn forbidden_funcs(&self) -> &BTreeSet<&'static str>

Return the set of functions that are forbidden to be called.

fn counters(&self) -> (u64, u64)

Return gas and gas allowance left in the counters.

fn update_counters(&mut self, gas: u64, allowance: u64)

Update counters with the provided values.

fn out_of_gas(&mut self) -> Self::Error

Handler for the case when gas is out.

fn out_of_allowance(&mut self) -> Self::Error

Handler for the case when gas allowance is out.

Provided Methods

fn send(
&mut self,
msg: HandlePacket,
delay: u32
) -> Result<MessageId, Self::Error>

Send message to another program.

fn reply(
&mut self,
msg: ReplyPacket,
delay: u32
) -> Result<MessageId, Self::Error>

Produce reply to the current message.