Trait gear_core::env::Ext

pub trait Ext {

    fn alloc(&mut self, pages: PageNumber) -> Result<PageNumber, &'static str>;

    fn block_height(&self) -> u32;

    fn block_timestamp(&self) -> u64;

    fn send_init(&mut self) -> Result<usize, &'static str>;

    fn send_push(
        &mut self, 
        handle: usize, 
        buffer: &[u8]
    ) -> Result<(), &'static str>;

    fn reply_commit(
        &mut self, 
        msg: ReplyPacket
    ) -> Result<MessageId, &'static str>;

    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), &'static str>;

    fn send_commit(
        &mut self, 
        handle: usize, 
        msg: OutgoingPacket
    ) -> Result<MessageId, &'static str>;

    fn reply_to(&self) -> Option<(MessageId, ExitCode)>;

    fn source(&mut self) -> ProgramId;

    fn message_id(&mut self) -> MessageId;

    fn free(&mut self, ptr: PageNumber) -> Result<(), &'static str>;

    fn debug(&mut self, data: &str) -> Result<(), &'static str>;

    fn set_mem(&mut self, ptr: usize, val: &[u8]);

    fn get_mem(&self, ptr: usize, buffer: &mut [u8]);

    fn msg(&mut self) -> &[u8];

    fn charge_gas(&mut self, amount: u32) -> Result<(), &'static str>;

    fn refund_gas(&mut self, amount: u32) -> Result<(), &'static str>;

    fn gas_available(&mut self) -> u64;

    fn value(&self) -> u128;

    fn wait(&mut self) -> Result<(), &'static str>;

    fn wake(&mut self, waker_id: MessageId) -> Result<(), &'static str>;

    fn send(&mut self, msg: OutgoingPacket) -> Result<MessageId, &'static str> { ... }

    fn reply(&mut self, msg: ReplyPacket) -> Result<MessageId, &'static str> { ... }

}

External api for managing memory, messages, allocations and gas-counting.

Required methods

fn alloc(&mut self, pages: PageNumber) -> Result<PageNumber, &'static str>

Allocate number of pages.

The resulting page number should point to pages consecutives memory pages.

fn block_height(&self) -> u32

Get the current block height.

fn block_timestamp(&self) -> u64

Get the current block timestamp.

fn send_init(&mut self) -> Result<usize, &'static str>

Initialize a new incomplete message for another program and return its handle.

fn send_push(
&mut self,
handle: usize,
buffer: &[u8]
) -> Result<(), &'static str>

Push an extra buffer into message payload by handle.

fn reply_commit(&mut self, msg: ReplyPacket) -> Result<MessageId, &'static str>

Complete reply message and send it to source program.

fn reply_push(&mut self, buffer: &[u8]) -> Result<(), &'static str>

Push an extra buffer into reply message.

fn send_commit(
&mut self,
handle: usize,
msg: OutgoingPacket
) -> Result<MessageId, &'static str>

Complete message and send it to another program.

fn reply_to(&self) -> Option<(MessageId, ExitCode)>

Read the message id, if current message is a reply.

fn source(&mut self) -> ProgramId

Get the source of the message currently being handled.

fn message_id(&mut self) -> MessageId

Get the id of the message currently being handled.

fn free(&mut self, ptr: PageNumber) -> Result<(), &'static str>

Free specific memory page.

Unlike traditional allocator, if multiple pages allocated via alloc, all pages should be free-d separately.

fn debug(&mut self, data: &str) -> Result<(), &'static str>

Send debug message.

This should be no-op in release builds.

fn set_mem(&mut self, ptr: usize, val: &[u8])

Set memory region at specific pointer.

fn get_mem(&self, ptr: usize, buffer: &mut [u8])

Reads memory contents at the given offset into a buffer.

fn msg(&mut self) -> &[u8]

Access currently handled message payload.

fn charge_gas(&mut self, amount: u32) -> Result<(), &'static str>

Charge some gas.

fn refund_gas(&mut self, amount: u32) -> Result<(), &'static str>

Refund some gas.

fn gas_available(&mut self) -> u64

Tell how much gas is left in running context.

fn value(&self) -> u128

Value associated with message.

fn wait(&mut self) -> Result<(), &'static str>

Interrupt the program and reschedule execution.

fn wake(&mut self, waker_id: MessageId) -> Result<(), &'static str>

Wake the waiting message and move it to the processing queue.

Provided methods

fn send(&mut self, msg: OutgoingPacket) -> Result<MessageId, &'static str>

Send message to another program.

fn reply(&mut self, msg: ReplyPacket) -> Result<MessageId, &'static str>

Produce reply to the current message.