Module gear_core::gas

Gas module.

Structs

GasAllowanceCounter

Gas allowance counter with some predefined maximum value.

GasAmount

Read-only representation of consumed GasCounter.

GasCounter

Gas counter with some predefined maximum gas.

GasLeft

Gas limit and gas allowance left.

ValueCounter

Value counter with some predefined maximum value.

Enums

ChargeError

Charging error

ChargeResult

The result of charging gas.

Traits

CountersOwner

Counters owner can change gas limit and allowance counters.

Token

This trait represents a token that can be used for charging GasCounter.