Trait gear_core::gas::Token

pub trait Token: Copy + Clone + Into<u64> {
    // Required method
    fn weight(&self) -> u64;
}

This trait represents a token that can be used for charging GasCounter.

Implementing type is expected to be super lightweight hence Copy (Clone is added for consistency). If inlined there should be no observable difference compared to a hand-written code.

Required Methods

fn weight(&self) -> u64

Return the amount of gas that should be taken by this token.

This function should be really lightweight and must not fail. It is not expected that implementors will query the storage or do any kinds of heavy operations.

That said, implementors of this function still can run into overflows while calculating the amount. In this case it is ok to use saturating operations since on overflow they will return max_value which should consume all gas.

Object Safety

This trait is not object safe.