Module gear_core::ids

Base identifiers for messaging primitives.

Modules

prelude

The purpose of this module is to make it easier to import gprimitives extensions.

Structs

ActorId

Program (actor) identifier.

CodeId

Code identifier.

MessageId

Message identifier.

ReservationId

Reservation identifier.

Functions

hash

Creates a unique identifier by passing given argument to blake2b hash-function.

hash_of_array

Creates a unique identifier by passing given argument to blake2b hash-function.

Type Aliases

ProgramId

Program identifier.