gear_core
Module memory

Module for memory and allocations context.

Structs

AllocationsContext

Pages allocations context for the running program.

IntoPageBufError

Error in attempt to make wrong size page buffer.

MemoryInterval

Interval in wasm program memory.

NoopGrowHandler

Grow handler do nothing implementation

PageBuf

Buffer for gear page data.

Enums

AllocError

Allocation error

MemoryError

Core memory error.

MemorySetupError

Inconsistency in memory parameters provided for wasm execution.

Traits

GrowHandler

Before and after memory grow actions.

Memory

Backend wasm memory interface.

Type Aliases

HostPointer

Host pointer type. Host pointer can be 64bit or less, to support both we use u64.

PageBufInner

Alias for inner type of page buffer.