Module gear_core::memory

Module for memory and allocations context.

Structs

AllocationsContext

Pages allocations context for the running program.

GrowHandlerNothing

Grow handler do nothing implementation

PageBuf

Buffer for gear page data.

PageNumber

Page number.

WasmPageNumber

Wasm page number.

Enums

Error

Memory error.

Constants

PAGE_STORAGE_GRANULARITY

Pages data storage granularity (PSG) is a size and wasm addr alignment of a memory interval, for which the following conditions must be met: if some gear page has data in storage, then all gear pages, that are in the same granularity interval, must contain data in storage. For example:

Traits

GrowHandler

Before and after memory grow actions.

Memory

Backend wasm memory interface.

Functions

vec_page_data_map_to_page_buf_map

Tries to convert vector data map to page buffer data map. Makes it without buffer reallocations.

Type Definitions

HostPointer

Host pointer type. Host pointer can be 64bit or less, to support both we use u64.