Module gear_core::memory

Module for memory and allocations context.

Structs

AllocInfo

Alloc method result.

AllocationsContext

Pages allocations context for the running program.

GearPage

Page number.

GranularityPage

Page with size PAGE_STORAGE_GRANULARITY.

IncorrectAllocationDataError

Incorrect allocation data error

MemoryInterval

Interval in wasm program memory.

NoopGrowHandler

Grow handler do nothing implementation

PageBuf

Buffer for gear page data.

PagesIter

U32 size pages iterator, to iterate continuously from one page to another.

PagesIterInclusive

U32 size pages iterator, to iterate continuously from one page to another, including the last one.

WasmPage

Wasm page number.

Enums

AllocError

Allocation error

PageError

Errors when act with PageU32Size.

Constants

GEAR_PAGE_SIZE

A gear page size, currently 4KiB to fit the most common native page size. This is size of memory data pages in storage. So, in lazy-pages, when program tries to access some memory interval - we can download just some number of gear pages instead of whole wasm page. The number of small pages, which must be downloaded, is depends on host native page size, so can vary.

PAGE_STORAGE_GRANULARITY

Pages data storage granularity (PSG) is a size and wasm addr alignment of a memory interval, for which the following conditions must be met: if some gear page has data in storage, then all gear pages, that are in the same granularity interval, must contain data in storage. For example:

WASM_PAGE_SIZE

A WebAssembly page has a constant size of 64KiB.

Traits

GrowHandler

Before and after memory grow actions.

Memory

Backend wasm memory interface.

PageU32Size

Trait represents page with u32 size for u32 memory: max memory size is 2^32 bytes. All operations with page guarantees, that no addr or page number can be overflowed.

Type Definitions

HostPointer

Host pointer type. Host pointer can be 64bit or less, to support both we use u64.

PageBufInner

Alias for inner type of page buffer.