gear_core::memory

Trait Memory

pub trait Memory<Context> {
    type GrowError: Debug;

    // Required methods
    fn grow(
        &self,
        ctx: &mut Context,
        pages: WasmPagesAmount,
    ) -> Result<(), Self::GrowError>;
    fn size(&self, ctx: &Context) -> WasmPagesAmount;
    fn write(
        &self,
        ctx: &mut Context,
        offset: u32,
        buffer: &[u8],
    ) -> Result<(), MemoryError>;
    fn read(
        &self,
        ctx: &Context,
        offset: u32,
        buffer: &mut [u8],
    ) -> Result<(), MemoryError>;
    unsafe fn get_buffer_host_addr_unsafe(&self, ctx: &Context) -> HostPointer;

    // Provided method
    fn get_buffer_host_addr(&self, ctx: &Context) -> Option<HostPointer> { ... }
}

Backend wasm memory interface.

Required Associated Types

type GrowError: Debug

Memory grow error.

Required Methods

fn grow( &self, ctx: &mut Context, pages: WasmPagesAmount, ) -> Result<(), Self::GrowError>

Grow memory by number of pages.

fn size(&self, ctx: &Context) -> WasmPagesAmount

Return current size of the memory.

fn write( &self, ctx: &mut Context, offset: u32, buffer: &[u8], ) -> Result<(), MemoryError>

Set memory region at specific pointer.

fn read( &self, ctx: &Context, offset: u32, buffer: &mut [u8], ) -> Result<(), MemoryError>

Reads memory contents at the given offset into a buffer.

unsafe fn get_buffer_host_addr_unsafe(&self, ctx: &Context) -> HostPointer

Get buffer addr unsafe.

Safety

If memory size is 0 then buffer addr can be garbage

Provided Methods

fn get_buffer_host_addr(&self, ctx: &Context) -> Option<HostPointer>

Returns native addr of wasm memory buffer in wasm executor