Trait gear_core::memory::Memory

pub trait Memory: Any {
    fn grow(&self, pages: PageNumber) -> Result<PageNumber, Error>;

    fn size(&self) -> PageNumber;

    fn write(&self, offset: usize, buffer: &[u8]) -> Result<(), Error>;

    fn read(&self, offset: usize, buffer: &mut [u8]);

    fn data_size(&self) -> usize;

    fn data_ptr(&self) -> *mut u8;

    fn clone(&self) -> Box<dyn Memory>;

    fn as_any(&self) -> &dyn Any;

    fn get_wasm_memory_begin_addr(&self) -> usize;

    fn set_pages(
        &self, 
        pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>
    ) -> Result<(), Error> { ... }
}

Memory interface for the allocator.

Required methods

fn grow(&self, pages: PageNumber) -> Result<PageNumber, Error>

Grow memory by number of pages.

fn size(&self) -> PageNumber

Return current size of the memory.

fn write(&self, offset: usize, buffer: &[u8]) -> Result<(), Error>

Set memory region at specific pointer.

fn read(&self, offset: usize, buffer: &mut [u8])

Reads memory contents at the given offset into a buffer.

fn data_size(&self) -> usize

Returns the byte length of this memory.

fn data_ptr(&self) -> *mut u8

Returns the base pointer, in the host’s address space, that the memory is located at.

fn clone(&self) -> Box<dyn Memory>

Clone this memory.

fn as_any(&self) -> &dyn Any

Downcast to exact memory type

fn get_wasm_memory_begin_addr(&self) -> usize

Returns native addr of wasm memory buffer in wasm executor

Provided methods

fn set_pages(
&self,
pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>
) -> Result<(), Error>

Set memory pages from PageBuf map, grow if possible.