Module gear_core::message

Message processing module.

Structs

ContextOutcome

Context outcome.

ContextOutcomeDrain

Context outcome dispatches and awakening ids.

ContextSettings

Context settings.

ContextStore

Store of previous message execution context.

Dispatch

Message with entry point.

HandleMessage

Message for Handle entry point. Represents a standard message that sends between actors.

HandlePacket

Handle message packet.

IncomingDispatch

Incoming message with entry point and previous execution context, if exists.

IncomingMessage

Incoming message.

InitMessage

Message for Init entry point. Used to initiate a newly created program.

InitPacket

Init message packet.

Message

An entity that is used for interaction between actors. Can transfer value and executes by programs in corresponding function: init, handle or handle_reply.

MessageContext

Context of currently processing incoming message.

PayloadSizeError

Payload size exceed error

ReplyDetails

Reply details data.

ReplyInfo

The struct contains results of read only send message RPC call.

ReplyMessage

Message for Reply entry point. ReplyMessage is unique because of storing MessageId from message on what it replies, and can be the only one per some message execution.

ReplyPacket

Reply message packet.

SignalDetails

Signal details data.

SignalMessage

Message for signal entry point.

StoredDelayedDispatch

Stored message with entry point.

StoredDispatch

Stored message with entry point and previous execution context, if exists.

StoredMessage

Stored message.

UserMessage

Message sent to user and deposited as event.

UserStoredMessage

Message sent to user and added to mailbox.

Enums

DispatchKind

Entry point for dispatch processing.

MessageDetails

Message details data.

MessageWaitedType

Composite wait type for messages waiting.

Constants

MAX_PAYLOAD_SIZE

Max payload size which one message can have (8 MiB).

Traits

Packet

Message packet.

WasmEntryPoint

Trait defining type could be used as entry point for a wasm module.

Type Aliases

GasLimit

Gas limit type for message.

Payload

Payload type for message.

Salt

Salt type for init message.

Value

Value type for message.