gear_core
Module pages

Module for memory pages.

Re-exports

pub use numerated;

pub use numerated::num_traits;

Structs

Page

Struct represents memory page number with some constant size SIZE in bytes.

PageError

Try from u32 error for Page.

PagesAmount

Struct represents memory pages amount with some constant size SIZE in bytes.

PagesAmountError

Try from u32 error for PagesAmount.

Type Aliases

GearPage

Page of gear page size - 16 kiB.

GearPagesAmount

Pages amount for gear page size - 16 kiB.

WasmPage

Page of wasm page size - 64 kiB.

WasmPagesAmount

Pages amount for wasm page size - 64 kiB.